WIP debugging

+++ source/kylarLEDs/EffectEngine/EffectEngine.h
@@ -12,12 +12,12 @@ class EffectEngine{
         EffectEngine();
         void run();
         void clear();
-        void clearPattern(uint8_t pattern_id);
+        void clearPattern(uint32_t pattern_id);
         void apply(Effect *effect);
         void queueApply(Effect* effect);
-        void setActivePattern(uint8_t patternIndex);
+        void setActivePattern(uint32_t patternIndex);
         std::vector<Effect *> effects;
 	    std::vector<Effect *> effectsQueue;
     private:
-        uint8_t activePattern;
+        uint32_t activePattern;
 };
